add TS for more components
